- Fixed small 3D view rendering problem with the far distance wall - Cheats settings are now saved in your user profile, so you don't have to keep turning them on - Time now passes correctly when resting or searching and after combat - Altered the Area Map to now overlay the arrow icon cleanly, and doors are shown in blue

+++ engine/ovr018.cs
@@ -178,7 +178,7 @@ namespace engine
 
                 if (var_3 == true)
                 {
-                    if (unk_4C13D.MemberOf((byte)inputkey) == true)
+                    if (unk_4C13D.MemberOf(inputkey) == true)
                     {
                         if (ovr026.is_human(gbl.player_ptr) == false ||
                             ovr026.getExtraFirstSkill(gbl.player_ptr) != 0x11)
@@ -216,7 +216,7 @@ namespace engine
                 }
                 else
                 {
-                    if (unk_4C15D.MemberOf((byte)inputkey) == false)
+                    if (unk_4C15D.MemberOf(inputkey) == false)
                     {
                         gbl.byte_1C01B = 0;
                     }
@@ -1058,7 +1058,7 @@ namespace engine
                             }
 
                             if (var_53.stats[var_1B].max < 13 &&
-                                asc_4C707.MemberOf((byte)var_53._class) == true)
+                                asc_4C707.MemberOf((int)var_53._class) == true)
                             {
                                 var_53.stats[var_1B].max = 13;
                             }
